TP2/src/requirements.txt
Pillow
PyQt5
PyQt5-sip
opencv-python